+++ lib/PrintAsObjC/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_swift_host_library(swiftPrintAsObjC STATIC
   DeclAndTypePrinter.cpp
+  ModuleContentsWriter.cpp
   PrintAsObjC.cpp)
 target_link_libraries(swiftPrintAsObjC PRIVATE
   swiftAST
